move IngredientCard to burger-ingredients folder and add CardOptions
