+++ src/components/burger-ingredients/index.ts
@@ -1,5 +1,6 @@
 import BurgerIngredients from "./burger-ingredients"
 
 export default BurgerIngredients
-export { IngredientsTabPanel } from "./ingredients-tab-panel"
+export { IngredientsTabPanel, type IngredientsTabPanelProps } from "./ingredients-tab-panel"
+export { IngredientCard, type IngredientCardProps } from "./ingredient-card"
 export type { BurgerIngredientsProps } from "./burger-ingredients"
