+++ src/components/common/card/index.ts
@@ -1,2 +1,1 @@
-export { IngredientCard } from "./ingredient-card"
 export { Card } from "./card"
